requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
aiosqlite==0.22.0
annotated-types==0.7.0
anyio==4.12.0
asyncpg==0.31.0
attrs==25.4.0
banks==2.2.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.18
dirtyjson==1.0.8
diskcache==5.6.3
distro==1.9.0
filelock==3.20.1
filetype==1.2.0
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.3.0
griffe==1.15.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.10
llama-index-cli==0.5.3
llama-index-core==0.14.10
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.2
llama-index-llms-llama-cpp==0.5.1
llama-index-llms-openai==0.6.12
llama-index-readers-file==0.5.5
llama-index-readers-llama-parse==0.5.1
llama-index-vector-stores-postgres==0.7.2
llama-index-workflows==2.11.5
llama-parse==0.6.54
llama_cpp_python==0.3.16
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.6.1
nltk==3.9.2
numpy==2.3.5
openai==2.13.0
packaging==25.0
pandas==2.2.3
pgvector==0.4.2
pillow==12.0.0
platformdirs==4.5.1
propcache==0.4.1
psutil==7.1.3
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
pypdf==6.4.2
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.45
striprtf==0.0.26
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
torch==2.9.1
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
wrapt==1.17.3
yarl==1.22.0

# Web UI
streamlit>=1.28.0
plotly>=5.18.0
umap-learn>=0.5.4

# MLX (Apple Silicon optimization - optional but recommended for M1/M2/M3 Macs)
mlx>=0.20.0
mlx-embedding-models>=0.0.11

# Advanced retrieval features (optional)
rank-bm25>=0.2.2
